Clase fin crud

+++ src/main/java/com/ar/bankingonline/application/services/UserService.java
@@ -1,13 +1,15 @@
 package com.ar.bankingonline.application.services;
 
-import com.ar.bankingonline.api.controllers.mappers.UserMapper;
+import com.ar.bankingonline.api.mappers.UserMapper;
+import com.ar.bankingonline.domain.exceptions.AccountNotFoundException;
 import com.ar.bankingonline.domain.models.User;
-import com.ar.bankingonline.api.controllers.dtos.UserDto;
+import com.ar.bankingonline.api.dtos.UserDto;
 import com.ar.bankingonline.infrastructure.repositories.UserRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
+import java.util.Optional;
 
 @Service
 public class UserService {
@@ -30,7 +32,7 @@ public class UserService {
     }
 
     // TODO: Refactor
-    public UserDto getUserById(Integer id){
+    public UserDto getUserById(Long id){
         return UserMapper.userMapToDto(repository.findById(id).get());
     }
 
@@ -39,12 +41,32 @@ public class UserService {
         return UserMapper.userMapToDto(repository.save(UserMapper.dtoToUser(user)));
     }
 
-    public UserDto update(UserDto user){
-        return UserMapper.userMapToDto(repository.save(UserMapper.dtoToUser(user)));
+    public UserDto update(Long id, UserDto user){
+
+        Optional<User> userCreated = repository.findById(id);
+
+        if (userCreated.isPresent()){
+            User entity = userCreated.get();
+
+            User accountUpdated = UserMapper.dtoToUser(user);
+
+            accountUpdated.setId(entity.getId());
+
+            User saved = repository.save(accountUpdated);
+
+            return UserMapper.userMapToDto(saved);
+        } else {
+            throw new AccountNotFoundException("User not found with id: " + id);
+        }
     }
 
-    public void delete(Integer id){
-        repository.deleteById(id);
+    public String delete(Long id){
+        if (repository.existsById(id)){
+            repository.deleteById(id);
+            return "Se ha eliminado la cuenta";
+        } else {
+            return "No se ha eliminado la cuenta";
+        }
     }
